move form creation functionality to fileHelper
