+++ src/modules/data/common-cartridge/common-cartridge-export.ts
@@ -1,3 +1,5 @@
+import { enforceDownload } from "@/utils/fileHelper";
+
 export type CommonCartridgeVersion = "1.1.0" | "1.3.0";
 
 export const allowedVersions: CommonCartridgeVersion[] = ["1.1.0", "1.3.0"];
@@ -13,29 +15,9 @@ export const startExport = async (
 		throw new Error(`Invalid version: ${version}. Allowed versions are: ${allowedVersions.join(", ")}`);
 	}
 
-	const createHiddenInputElement = (name: string, value: string): HTMLInputElement => {
-		const input = document.createElement("input");
-		input.type = "hidden";
-		input.name = name;
-		input.value = value;
-		return input;
-	};
-
-	const form = document.createElement("form");
-	form.method = "POST";
-	form.action = `/api/v3/common-cartridge/export/${roomId}?version=${version}`;
-	form.enctype = "application/json";
-	form.target = "_blank";
-
-	const topicIdsInput = createHiddenInputElement("topics", JSON.stringify(topics));
-	const taskIdsInput = createHiddenInputElement("tasks", JSON.stringify(tasks));
-	const columnBoardIdsInput = createHiddenInputElement("columnBoards", JSON.stringify(columnBoards));
-
-	form.appendChild(topicIdsInput);
-	form.appendChild(taskIdsInput);
-	form.appendChild(columnBoardIdsInput);
-
-	document.body.appendChild(form);
-	form.submit();
-	document.body.removeChild(form);
+	enforceDownload(`/api/v3/common-cartridge/export/${roomId}?version=${version}`, [
+		{ key: "topics", value: JSON.stringify(topics) },
+		{ key: "tasks", value: JSON.stringify(tasks) },
+		{ key: "columnBoards", value: JSON.stringify(columnBoards) },
+	]);
 };
